Customize request routing outside Knots: clean up
