+++ knotx-gateway/pom.xml
@@ -26,9 +26,9 @@
   </parent>
   <modelVersion>4.0.0</modelVersion>
 
-  <artifactId>knotx-knot-custom-flow</artifactId>
+  <artifactId>knotx-gateway</artifactId>
 
-  <name>Knot.x - Reactive microservice assembler - Knot - Custom Flow</name>
+  <name>Knot.x - Reactive microservice assembler - Knot - Gateway</name>
 
   <dependencies>
     <dependency>
