+++ knotx-gateway/src/main/java/io/knotx/gateway/configuration/CustomFlowKnotConfiguration.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.knotx.knot.customflow.configuration;
+package io.knotx.gateway.configuration;
 
 import io.vertx.core.json.JsonObject;
