+++ knotx-gateway/src/main/java/io/knotx/gateway/impl/ResponseProviderKnotProxyImpl.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.knotx.knot.customflow.impl;
+package io.knotx.gateway.impl;
 
 import java.util.NoSuchElementException;
 import java.util.Set;
