+++ knotx-gateway/src/main/resources/io.knotx.GatewayKnot.json
@@ -2,7 +2,7 @@
   "main": "io.knotx.knot.customflow.GatewayKnotVerticle",
   "options": {
     "config": {
-      "address": "knotx.knot.gatewayknot"
+      "address": "knotx.gatewayknot"
     }
   }
 }
