+++ knotx-gateway/src/main/resources/io.knotx.RequestProcessorKnot.json
@@ -1,0 +1,8 @@
+{
+  "main": "io.knotx.gateway.RequestProcessorKnotVerticle",
+  "options": {
+    "config": {
+      "address": "knotx.requestprocessor"
+    }
+  }
+}
