+++ knotx-gateway/src/main/resources/io.knotx.ResponseProviderKnot.json
@@ -1,0 +1,8 @@
+{
+  "main": "io.knotx.gateway.ResponseProviderKnotVerticle",
+  "options": {
+    "config": {
+      "address": "knotx.responseprovider"
+    }
+  }
+}
