+++ knotx-gateway/src/main/resources/knotx-gateway.json
