--- knotx-knot/knotx-knot-custom-flow/src/main/resources/io.knotx.ResponseProviderKnot.json
@@ -1,8 +1,0 @@
-{
-  "main": "io.knotx.knot.customflow.ResponseProviderKnotVerticle",
-  "options": {
-    "config": {
-      "address": "knotx.knot.responseprovider"
-    }
-  }
-}
